reBuild/pii/requirements.txt
boto3
paddlepaddle
paddleocr
pymupdf
pillow
numpy
opencv-python-headless
psycopg2-binary
python-dotenv
